Debugged + Added Comments + Call Mini for Black
Debugging revealed that the ai gets stuck on move repition.
The probable reason is that it doesnt look past the maximum depth, and hence it may find a favorable position upto that depth but the next of blacks move maybe upset the favor.
Hence in ai vs ai they get stuck in a loop as noone looks past the max depth
Another reason is that a rule for 3-fold repition = draw is not implemented

Good fix is quiuescience
Lazy fix is repition  = bad but it has its own bugs