Ok less pain now. Please read desc
So many of the bugs arose from the fact that you dont look further than the given depth, causing it to make small peice blunders such as pawns or knights.

The fix and main highlight of this commit: Give a less eval when its the opponents turn , and the more things he can capture the less that eval is, meaning it discourages moves where alot of your peices can be taken at the end. This is a very dumb rule, plan to change it to take into account the cost o captures and possibly a capture only search.

In other news, migrated the stalemates   checkmates etc check to eval function. reduced weights a bit coz idk felt like it
In last commit, i had changed the equal eval check to take into account "attack periods" of the game. I had previously only given attacking moves priority (but accidentally gave them less priority in mini() ). Now in both mini and max, attacking move sonly have better priority if they are done during the midgame and otherwise, only moves which take something greater than pawns have a right to be prioritized.
Further, c++ double has issues, it makes 2.9999999999999999999 and 3.0000000000000001 and this causes the two evals to not be equal, so introduced a little error check thingy which entertains a difference of 0.01 in evals and deems them to be equal.
Also orderMoves() also returns the number of captures moves now

Also reduced the doubled pawn evals to 0.1. Endgame now starts when material count is <= 40 (meaning 2 queens 2 rooks a knight/bishop and few pawns.. i guess ill make it 30 if need be